move draft views into their own folder
