+++ packages/front-end/src/views/draft/DraftPoolView.scss
