+++ packages/front-end/src/views/draft/LiveDraftPoolView.scss
